Move internal classes to internal package.
This refactoring is help setup MiMa binary compatibility checks.
There are no binary compatibility promises for code under the scalafix.internal
package, while code outside the internal package will remain binary
compatible across each 0.x series.
