+++ scalafix-cli/src/main/scala/scalafix/internal/cli/WriteMode.scala
@@ -1,4 +1,4 @@
-package scalafix.cli
+package scalafix.internal.cli
 
 sealed abstract class WriteMode {
   def isWriteFile: Boolean = this == WriteMode.WriteFile
