+++ scalafix-core/shared/src/main/scala/scalafix/internal/util/FileOps.scala
@@ -1,4 +1,4 @@
-package scalafix.util
+package scalafix.internal.util
 
 import java.io.BufferedReader
 import java.io.File
