+++ scalafix-core/shared/src/main/scala/scalafix/internal/util/TokenOps.scala
@@ -1,4 +1,4 @@
-package scalafix.util
+package scalafix.internal.util
 
 import scala.meta.Token
