--- scalafix-core/shared/src/main/scala/scalafix/util/package.scala
@@ -1,3 +1,0 @@
-package scalafix
-
-package object util {}
